[Refactor] File location change
- 리팩토링 작업.
- 파일 구조화.
