+++ src/pages/RestaurantPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useParams } from 'react-router-dom';
 
-import RestaurantContainer from './RestaurantContainer';
+import RestaurantContainer from '../containers/RestaurantContainer';
 
 export default function RestaurantPage({ params }) {
   const { id } = params || useParams();
