+++ src/store/slice.test.js
@@ -272,7 +272,7 @@ describe('reducer', () => {
 const middlewares = getDefaultMiddleware();
 const mockStore = configureStore(middlewares);
 
-jest.mock('./services/api');
+jest.mock('../services/api');
 
 describe('actions', () => {
   let store;
